namespace: include required_parts in routes helper
This commit exports the `required_parts` of routes that are exposed
for Administrate.

Instead of only exporting the defaults `controller` and `action`
names, we also export the `required_parts` (required parameters).

This can be useful when we want to check if a route is available for a
resource (and especially for a nested resource).