existing_action?: verify if a route is available for a nested model
This commit adapts the current `existing_action?` helper to be able to
pass another “parent resource” argument when we want to check if a
nested resource has a defined route.

It's usefull in the example where your model only makes sense when
it is accessed from a parent model.

E.g. In the following routes, `Order` model can only be shown (or
edited) when it's accessed via it's nested
`/customers/:customer_id/orders` route.
```
Rails.application.routes.draw do
  namespace(:admin) do
    resources :customers do
      resources :orders, only: [:show, :edit]
    end
  end
end
```