views(collection): use the new feature of `valid_action?`
This commit uses the new feature of ̀valid_action?` where we can check
if an action is available as a “nested action”. This makes sense when
models are nested within a parent related model and when we only want
those resources to be available in the context of their parent.

We thus generate correct URLs depending on which routes are defined.